+++ packages/studio-components/src/index.tsx
@@ -4,7 +4,6 @@ export { default as SegmentedTabs } from './SegmentedTabs';
 export { default as TableCard } from './TableCard';
 export { default as Toolbar } from './Toolbar';
 export { default as Logo, LogoImage, LogoText } from './Logo';
-export { default as Container } from './Container';
 export { default as PropertiesList } from './PropertiesList';
 export { default as MappingFields } from './PropertiesList/MappingFields';
 export { default as Section } from './Section';
@@ -17,7 +16,6 @@ export { default as SideTabs } from './SideTabs';
 export { default as ResizablePanels } from './ResizablePanel';
 export { default as FullScreen } from './FullScreen';
 export { default as Slot } from './Slot';
-export { default as ContentSection } from './ContentSection';
 export { default as TypingText } from './TypingText';
 export { default as CreatePortal } from './CreatePortal';
 export { default as Layout } from './layout';
@@ -26,7 +24,6 @@ export * as Utils from './Utils';
 export * as Icons from './Icons';
 
 /** export hooks  */
-export { useContainer } from './Container/useContainer';
 export { useSection } from './Section/useSection';
 export { useMultipleInstance } from './MultipleInstance/useMultipleInstance';
 export { useStudioProvier } from './Provider/useThemeConfigProvider';