+++ packages/studio-graph/src/components/index.tsx
@@ -9,7 +9,6 @@ export { default as StyleSetting } from './StyleSetting';
 export { default as Prepare } from './Prepare';
 export { default as ZoomFit } from './ZoomFit';
 export { default as ClearStatatus } from './ClearStatus';
-export { default as SegmentedTabs } from './SegmentedTabs';
 export { default as SliderFilter } from './SliderFilter';
 export { default as RunCluster } from './RunCluster';
 export { default as LayoutSetting } from './LayoutSetting';