+++ packages/studio-explore/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphscope/studio-explore
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphscope/studio-components@0.1.7
+  - @graphscope/studio-graph@0.1.7
+  - @graphscope/studio-importor@0.1.7
+
 ## 0.1.4
 
 ### Patch Changes