+++ packages/studio-importor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphscope/studio-importor
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphscope/studio-components@0.1.7
+
 ## 0.1.4
 
 ### Patch Changes